Spring quickstart: change package from com.example to org.acme (like all other quickstarts)
