+++ technology/java-spring-boot/build.gradle
@@ -6,7 +6,7 @@ plugins {
 
 def optaplannerVersion = "8.12.0-SNAPSHOT"
 
-group = "com.example"
+group = "org.acme"
 version = "0.1.0-SNAPSHOT"
 sourceCompatibility = "11"
