+++ technology/java-spring-boot/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.example</groupId>
+  <groupId>org.acme</groupId>
   <artifactId>optaplanner-spring-boot-school-timetabling-quickstart</artifactId>
   <version>1.0-SNAPSHOT</version>
