+++ technology/java-spring-boot/src/main/java/org/acme/schooltimetabling/domain/Room.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.schooltimetabling.domain;
+package org.acme.schooltimetabling.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
