+++ technology/java-spring-boot/src/main/java/org/acme/schooltimetabling/persistence/TimeTableRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package com.example.schooltimetabling.persistence;
+package org.acme.schooltimetabling.persistence;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.schooltimetabling.domain.Lesson;
-import com.example.schooltimetabling.domain.TimeTable;
+import org.acme.schooltimetabling.domain.Lesson;
+import org.acme.schooltimetabling.domain.TimeTable;
 
 @Service
 @Transactional
