+++ technology/java-spring-boot/src/main/java/org/acme/schooltimetabling/persistence/TimeslotRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package com.example.schooltimetabling.persistence;
+package org.acme.schooltimetabling.persistence;
 
 import java.util.List;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import com.example.schooltimetabling.domain.Timeslot;
+import org.acme.schooltimetabling.domain.Timeslot;
 
 public interface TimeslotRepository extends PagingAndSortingRepository<Timeslot, Long> {
