+++ technology/java-spring-boot/src/test/java/org/acme/schooltimetabling/solver/TimeTableBenchmarkTest.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.acme.schooltimetabling.solver;
+
+import org.acme.schooltimetabling.rest.TimeTableController;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Timeout;
+import org.optaplanner.benchmark.api.PlannerBenchmarkFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class TimeTableBenchmarkTest {
+
+    @Autowired
+    private TimeTableController timeTableController;
+
+    @Autowired
+    private PlannerBenchmarkFactory benchmarkFactory;
+
+    @Test
+    @Timeout(600_000)
+    public void benchmark() {
+        benchmarkFactory.buildPlannerBenchmark(timeTableController.getTimeTable()).benchmark();
+    }
+}
